+++ src/seeder.js
@@ -49,7 +49,7 @@ async function runSeeder() {
                 id SERIAL NOT NULL PRIMARY KEY,
                 user_id INT NOT NULL,
                 motorcycle_id INT NOT NULL,
-                type SMALLINT NOT NULL,
+                type INT NOT NULL,
                 request VARCHAR(1023),
                 service_time TIMESTAMP NOT NULL,
                 service_status INT NOT NULL,
